feat: 게시글 조회 API 구현

+++ assignment4/src/main/java/com/jian/community/domain/model/PostView.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
@@ -13,7 +13,7 @@ public class PostView extends MinimalEntity {
 
     @Getter
     private Long postId;
-    private AtomicLong count = new AtomicLong(0);
+    private AtomicInteger count = new AtomicInteger(0);
 
     public static PostView from(Long postId) {
         PostView postView = new PostView();
@@ -21,11 +21,11 @@ public class PostView extends MinimalEntity {
         return postView;
     }
 
-    public Long increase(){
+    public int increase(){
         return count.incrementAndGet();
     }
 
-    public Long getCount(){
+    public int getCount(){
         return count.get();
     }
 }
